separated/deduplicated routes & handlers - handlers in handlers.js - routes in routes.js
